+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 ### Added
 
 - Turtle decoder
+- `RDF.List` module with functions for working with RDF lists
 - `describes?/1` on `RDF.Data` protocol and all RDF data structures which check  
   if statements about a given resource exist
 - `RDF.Data.descriptions/1` which returns all descriptions within a RDF data structure